+++ taichi/codegen/opengl/codegen_opengl.h
@@ -2,7 +2,7 @@
 
 #include "taichi/inc/constants.h"
 #include "taichi/lang_util.h"
-#include "taichi/backends/opengl/struct_opengl.h"
+#include "taichi/codegen/opengl/struct_opengl.h"
 #include "taichi/runtime/opengl/opengl_api.h"
 
 #include <string>
