+++ taichi/program/program.cpp
@@ -9,7 +9,7 @@
 #include "taichi/struct/struct_llvm.h"
 #include "taichi/backends/metal/api.h"
 #include "taichi/backends/wasm/aot_module_builder_impl.h"
-#include "taichi/backends/opengl/opengl_program.h"
+#include "taichi/runtime/program_impls/opengl/opengl_program.h"
 #include "taichi/backends/metal/metal_program.h"
 #include "taichi/backends/cc/cc_program.h"
 #include "taichi/platform/cuda/detect_cuda.h"
