+++ taichi/runtime/opengl/opengl_api.cpp
@@ -2,8 +2,8 @@
 
 #include <list>
 
-#include "taichi/backends/opengl/opengl_kernel_util.h"
-#include "taichi/backends/opengl/opengl_utils.h"
+#include "taichi/runtime/opengl/opengl_kernel_util.h"
+#include "taichi/runtime/opengl/opengl_utils.h"
 #include "taichi/runtime/opengl/shaders/runtime.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/program/kernel.h"
