+++ taichi/runtime/opengl/opengl_api.h
@@ -6,7 +6,7 @@
 
 #include "taichi/backends/device.h"
 #include "taichi/runtime/opengl/opengl_kernel_launcher.h"
-#include "taichi/backends/opengl/opengl_kernel_util.h"
+#include "taichi/runtime/opengl/opengl_kernel_util.h"
 #include "taichi/common/core.h"
 #include "taichi/ir/offloaded_task_type.h"
 #include "taichi/ir/transforms.h"
