+++ taichi/runtime/program_impls/CMakeLists.txt
@@ -1,3 +1,4 @@
 # ./taichi/runtime/program_impls/CMakeLists.txt
 
 add_subdirectory(vulkan)
+add_subdirectory(opengl)
